Update frontend application env file attributes
- env.development to store local development related values
