+++ reading-list-front-end/.env.development
@@ -1,0 +1,4 @@
+VITE_REDIRECT_URL=http://localhost:5173
+VITE_ASGARDEO_CLIENT_ID=
+VITE_ASGARDEO_BASE_URL=
+VITE_CHOREO_BACKEND_URL=
