--- reading-list-front-end/.env.example
@@ -1,9 +1,0 @@
-VITE_SIGNIN_REDIRECT_URL=
-VITE_SIGNOUT_REDIRECT_URL=
-VITE_ASG_CLIENT_ID=
-VITE_BASE_URL=
-VITE_RESOURCE_HOST=
-VITE_RESOURCE_SERVICE_NAME=
-VITE_RESOURCE_VERSION=
-VITE_RESOURCE_HANDLE=
-VITE_RESOURCE_ENDPOINT_CONTEXT=
