+++ reading-list-front-end/.gitignore
@@ -12,6 +12,7 @@ dist
 dist-ssr
 *.local
 .env
+.env.development
 
 # Editor directories and files
 .vscode/*
