--- reading-list-front-end/src/api/create-base-url.ts
@@ -1,14 +1,0 @@
-export const createBaseURL = ({
-  host,
-  service,
-  version,
-  handle,
-  endpointContext,
-}: {
-  host: string;
-  env: string;
-  service: string;
-  version: string;
-  handle: string;
-  endpointContext?: string;
-}) => `https://${host}/${handle}/${service}/${endpointContext}/${version}`;
